pyproject.toml
[project]
name = "diffengine"
version = "0.1.0"
description = "diffengine"
readme = "README.md"
keywords = [
  "computer vision",
  "diffusion models",
]
license = { file = "LICENSE" }
authors = [{ name = "takuoko", email = "to78314910@gmail.com" }]
requires-python = ">= 3.10"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "bs4",
  "clean-fid",
  "datasets>=2.19",
  "diffusers>=0.27.2",
  "ftfy",
  "joblib",
  "mmengine>=0.10.3",
  "open_clip_torch>=2.24",
  "peft>=0.10",
  "sentencepiece>=0.2",
  "torch>=2.1.2",
  "torch-optimi",
  "torchvision>=0.16.2",
  "tqdm",
  "transformers>=4.40",
  "ujson",
  "webdataset",
]
[project.optional-dependencies]
dev = [
  "coverage",
  "pre-commit",
  "pytest",
]
docs = [
  "docutils==0.18.1",
  "modelindex",
  "myst-parser",
  "pytorch_sphinx_theme@ git+https://github.com/mzr1996/pytorch_sphinx_theme",
  "sphinx==6.1.3",
  "sphinx-autoapi",
  "sphinx-autodoc-typehints",
  "sphinx-copybutton",
  "sphinx-notfound-page",
  "sphinx-rtd-theme",
  "sphinx-tabs",
  "sphinxcontrib-jquery",
  "tabulate",
]
optional = [
  "colossalai",
  "deepspeed",
  "hpsv2",
  "t2v-metrics",
]
[project.scripts]
diffengine = "diffengine:cli"

[tool.setuptools.packages.find]
exclude = [
    "build*",
    "docs*",
    "dist*",
    "tests*",
    "docker*",
    "data*",
    "work_dirs*",
    "projects*",
]

[tool.distutils.bdist_wheel]
universal = true

[tool.ruff]
target-version = "py310"
exclude = ["diffengine/patches/*"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "PGH004",
    "B028",
    "ANN002",
    "ANN003",
    "ERA001",
    "C408",
    "INP001",
    "D100",
    "D107",
    "D104",
    "S101",
    "ANN101",
    "PTH118",
    "S311",
]

[tool.ruff.lint.pylint]
max-args = 10

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "PLR2004",
    "N806",
    "PTH",
    "SLF001",
    "PT009",
    "D",
    "ANN201",
    "INP001",
    "ANN202",
    "ANN001",
]
"diffengine/tools/*" = [
    "T201", "ISC002", "PTH", "D", "INP001",
    "ANN201", "ANN202", "ANN001"]
"projects/*" = ["F403", "F405", "TID252"]
"diffengine/configs/*" = ["F403", "F405", "TID252"]

[tool.codespell]
ignore-words-list = "enver,"

[tool.mypy]
python_version = "3.11"
no_strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.wheel]
exclude = [
    "build*",
    "docs*",
    "dist*",
    "tests*",
    "docker*",
    "data*",
    "work_dirs*",
    "projects*",
]
